Auto-set the GLOBAL_AGENT_ENVIRONMENT_VARIABLE_NAMESPACE env var in the container image
Setting it to an empty value allows global-agent to rely on the rather
conventional HTTP(S)_PROXY and NO_PROXY env vars. [1]
Otherwise, users would need to configure both GLOBAL_AGENT_HTTP(S)_PROXY
and HTTP(S)_PROXY to make proxying work with both global-agent and other
libs.
Also, it might be a bit confusing for the user to have to set it explicitly
to an empty value.
We can set it the other way around, but users can still override it if
needed.

[1] https://github.com/gajus/global-agent#what-is-the-reason-global-agentbootstrap-does-not-use-http_proxy